refactor/config: re-partition config modules
